# express-api/assets/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vehicle Image API",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
    "/jobs": {
      "post": {
        "summary": "Queue one image; /process is an asynchronous alias",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "isBackgroundWant": {
                    "type": "boolean",
                    "default": false
                  },
                  "folder": {
                    "type": "string",
                    "enum": [
                      "parking-lots",
                      "backgrounds"
                    ],
                    "default": "parking-lots"
                  },
                  "background": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "default": "1.png"
                  },
                  "enhancement": {
                    "type": "boolean",
                    "default": false
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "imageurl"
                ],
                "properties": {
                  "isBackgroundWant": {
                    "type": "boolean",
                    "default": false
                  },
                  "folder": {
                    "type": "string",
                    "enum": [
                      "parking-lots",
                      "backgrounds"
                    ],
                    "default": "parking-lots"
                  },
                  "background": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "default": "1.png"
                  },
                  "enhancement": {
                    "type": "boolean",
                    "default": false
                  },
                  "imageurl": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted: jobId, statusUrl, resultUrl, pollAfterMs"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid API key"
          },
          "413": {
            "description": "Upload too large"
          },
          "429": {
            "description": "Queue full; Retry-After header"
          },
          "503": {
            "description": "Insufficient storage"
          }
        }
      }
    },
    "/process": {
      "post": {
        "summary": "Queue one image; /process is an asynchronous alias",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "isBackgroundWant": {
                    "type": "boolean",
                    "default": false
                  },
                  "folder": {
                    "type": "string",
                    "enum": [
                      "parking-lots",
                      "backgrounds"
                    ],
                    "default": "parking-lots"
                  },
                  "background": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "default": "1.png"
                  },
                  "enhancement": {
                    "type": "boolean",
                    "default": false
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "imageurl"
                ],
                "properties": {
                  "isBackgroundWant": {
                    "type": "boolean",
                    "default": false
                  },
                  "folder": {
                    "type": "string",
                    "enum": [
                      "parking-lots",
                      "backgrounds"
                    ],
                    "default": "parking-lots"
                  },
                  "background": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "default": "1.png"
                  },
                  "enhancement": {
                    "type": "boolean",
                    "default": false
                  },
                  "imageurl": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted: jobId, statusUrl, resultUrl, pollAfterMs"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid API key"
          },
          "413": {
            "description": "Upload too large"
          },
          "429": {
            "description": "Queue full; Retry-After header"
          },
          "503": {
            "description": "Insufficient storage"
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Poll queued, running, done, or failed status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status, stage, queuePosition, warning, error, resultUrl"
          },
          "404": {
            "description": "Unknown or expired"
          }
        }
      }
    },
    "/jobs/{id}/result": {
      "get": {
        "summary": "Download completed image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or expired"
          },
          "409": {
            "description": "Not ready"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "health",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": []
      }
    },
    "/ready": {
      "get": {
        "summary": "ready",
        "responses": {
          "200": {
            "description": "Success"
          },
          "503": {
            "description": "Worker not ready"
          }
        },
        "security": []
      }
    },
    "/backgrounds": {
      "get": {
        "summary": "backgrounds",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  }
}